Fix GCC compiler warning in readwholefileverb, introduced in previous commit.
git-svn-id: https://frontierkernel.svn.sourceforge.net/svnroot/frontierkernel/Frontier/trunk@1250 4e289971-d10f-0410-980f-25908192c0f7

+++ Common/source/fileverbs.c
@@ -2349,7 +2349,7 @@ static boolean readwholefileverb (hdltreenode hparam1, tyvaluerecord *v) {
 	if (!getpathvalue (hparam1, 1, &fs))
 		return (false);
 
-	if (!fifopenfile (&fs, currentprocess))
+	if (!fifopenfile (&fs, (long) currentprocess))
 		return (false);
 
 	fl = fifreadfile (&fs, &x);
